// mountain-traveler/front-end/src/app/common/constants/FiltersValues.enum.ts
export enum CommonValue {
    all = 'all'
}

export enum TripDurationValue {
    short = 'short',
    medium = 'medium',
    long = 'long'
};

export enum RouteTypeValue {
    oneWay = 'oneWay',
    roundTrip = 'roundTrip'
};

export enum HikingLevelsValue {
    easy = 'easy',
    medium = 'medium',
    hard = 'hard',
}